requirements.txt
Flask==0.12.1
wtforms
prompt_toolkit
azure-mgmt-resource~=1.1.0
azure-mgmt-network~=1.0.0
azure-mgmt-compute~=2.0.0
haikunator
azure-mgmt-storage
